… benchmarks/distributed/zero/analyze.py