@blefaudeux 2021-06-06
> > In the current ZeRO implementation, the core partition_parameters() computation happens twice upon initialization (i.e. call to **init**()):
> 
> @blefaudeux is this intentional? Is not, we can work on a followup PR to consolidate the two into one.

no it's not intentional, mistake indeed, good catch !