@blefaudeux 2021-06-06
Thanks for the very comprehensive write up @andwgu ! For completeness, I'll add a consideration to the partitioning: something else which could be done, and which I think could be meaningful speed-wise, is to make the partitioning order-aware, and DDP non-blocking on the reduce part in that case (each rank could just check that its own parameters are reduced before stepping, the lock being moved to after the step/broadcast). It's a much, much bigger change but one which allow for some overlap in between the backward and the step. Just food for thoughts, I think that we may have discussed that somewhere else already