@mrshenli 2021-06-07
Hey @andwgu, shall we move the benchmark code into a separate PR? If you need to have that PR sitting on top of this PR, you can use the ghstack tool: https://github.com/ezyang/ghstack

Thanks!