@pritamdamania87 2021-06-18
> Thanks a lot for conducting comprehensive experiments. Given that this new algorithm considerably improves the result for ResNet and only slightly degrades for BERT, I'd vote for landing this PR.

What is the high level metric we're trying to optimize for here? For example the optimizer.step() latency doesn't change much based on these changes. If we're looking at better load balancing of parameters for GPU memory utilization, we should probably track that metric and see if there is substantial improvement in balanced GPU memory. Also, I am concerned about the results for BERT since Transformer models are actually the number one use case where techniques like ZeRO optimizer would be used.