@pritamdamania87 2021-06-18
> however, given that n in typical use cases is very large, any algorithm that is quadratic or slower is unrealistic.

How much is n for the models we compared here? My understanding is that `n` here is not actually the total number of parameters in the model but `len(model.parameters())` which is probably much lower? I'm wondering if this algorithm is feasible: https://en.wikipedia.org/wiki/Multiway_number_partitioning#Reduction_to_bin_packing?