@andwgu 2021-06-18
> How much is n for the models we compared here? My understanding is that `n` here is not actually the total number of parameters in the model but `len(model.parameters())` which is probably much lower? I'm wondering if this algorithm is feasible: https://en.wikipedia.org/wiki/Multiway_number_partitioning#Reduction_to_bin_packing?

You are right, `n` is not the total number of parameters in the model but rather `len(model.parameters())`. My fault for any confusion. `len(model.parameters())` (i.e. `len(param_group["params"]` assuming one parameter group) for each of the three models are as follows:
- ResNet-50: 161
- ResNet-152: 467
- BERT: 199

Given the magnitude of these numbers, my intuition is that we can relax the time complexity requirement, but the constant factors will matter. We should also be clear on when this partitioning takes place: Right now, this is done once in the constructor and also any time that the parameter trainability changes (e.g. if some parameters are frozen or unfrozen). If we do not anticipate the parameter trainability to change often, then a more computationally intensive partition algorithm can be considered. Alternatively, we can use a heavier-weight algorithm at construction time and use faster algorithms thereafter, or we can allow the user to specify the algorithm(s) from a fixed list of choices (I say _fixed_ since I think that user-defined algorithms may be challenging to integrate).

I agree that we should establish what metric we are trying to optimize. Initially, I assumed the goodness of the partition would directly translate to an actual performance-impacting metric, but now I see that this may not be the case. As you mentioned, the optimizer `step()` latency is roughly the same. I am not experienced in the collective communications/networks, but perhaps, the usage of `broadcast()` in the `step()` implementation poses a limit? Each rank broadcasts its shard of the parameters asynchronously, but the degree to which a more-uniform partition helps depends on how much the broadcasts are overlapped.

Either way, I took a look at the linked algorithm. Denote multiway number partitioning (our problem) by MNP and bin packing by BP. To provide a brief summary: In both MNP and BP, we have `n` positive integers `S = {s_1, ..., s_n}`. In MNP, we want to assign them to `k` bins so that the largest bin is minimized. In BP, we have a capacity `C` and want to assign them into the minimum number of bins so that no bin exceeds `C`. The algorithm (solving MNP using a BP oracle) starts with a cleverly-chosen lower and upper bound on `C` and binary searches for the minimum `C` such that the `n` integers can be packed into at most `k` bins.

My concern is that finding the lower and upper bounds each take at least one pass over the parameters (or more precisely, at least `n + n log n` work for even the simplest algorithms). Then, if the difference in the bounds is `m`, there are another `log m` calls to the bin packing oracle, which itself requires at least one pass over the parameters. This results in several more passes over the parameters compared to the simple greedy algorithm or the sorted-greedy algorithm. Again, it is difficult to evaluate this tradeoff without first identifying the metric to optimize. 

If we decide that GPU memory utilization is the metric to track, let me know, and I can rework the experiment and update the results.