@andwgu 2021-06-21
Here, I present the results of the [new partitioning algorithm](https://en.wikipedia.org/wiki/Multiway_number_partitioning#Reduction_to_bin_packing) suggested by @pritamdamania.

**TL;DR: I incorrectly implemented the sorted-greedy algorithm previously, and the new algorithm performs no better than the sorted-greedy algorithm.**

**Algorithm:**
ZeRO's partitioning is an instance of the multiway number partitioning (MNP) problem: Suppose we have `n` positive integers `s_1, ..., s_n` and `k` bins. We want to pack elements into bins so that the largest bin is minimized. A dual to MNP is the bin packing (BP) problem: Given a max capacity `C` per bin, we want to pack the `n` elements into as few bins as possible. The new algorithm (sometimes known as multifit) leverages the duality by binary searching for the minimum capacity `C` such that the `n` elements can be packed into `k` or fewer bins. 

In my implementation, I lower bound the search space for `C` by `(sum_{i=1}^{n} s_i) / k` and upper bound it by `min{max bin size from greedy, max bin size from sorted-greedy}`. This upper bound implies that the new algorithm cannot do worse than the greedy and sorted-greedy algorithms. However, we shall see below that on the three considered models, the algorithm **only matches** the sorted-greedy algorithm's performance.

**Disclaimer:**
In the process of implementing this new algorithm, I caught a crucial mistake in my implementation in the sorted-greedy algorithm. I was sorting by the first dimension of each tensor's size instead of the number of elements in the tensor:
```
# Old incorrect version:
params_sorted = sorted(param_group["params"], key=lambda t: t.size()[0], reverse=True)
# Corrected version:
params_sorted = sorted(param_group["params"], key=lambda t: t.numel(), reverse=True)
```

**Updated Results:**
Two processes:
| Model | Max Num Params - Greedy (Diff) | Max Num Params - Sorted-Greedy (Diff) | Max Num Params - Multifit (Diff) | Mean Num Params |
| --- | --- | --- | --- | --- |
| ResNet-50 | 13,249,600 (471,084) | **12,778,536 (20)** | **12,778,536 (20)** | 12,778,516 |
| ResNet-152 | 30,567,488 (471,084) | **30,096,424 (20)** | **30,096,424 (20)** | 30,096,404 |
| BERT | 54,749,184 (8,064) | **54,741,504 (384)** | **54,741,504 (384)** | 54,741,120 |

Four processes:
| Model | Max Num Params - Greedy (Diff) | Max Num Params - Sorted-Greedy (Diff) | Max Num Params - Multifit (Diff) | Mean Num Params |
| --- | --- | --- | --- | --- |
| ResNet-50 | 7,524,864 (1,135,606) |  **6,389,288 (30)** | **6,389,288 (30)** | 6,389,258 |
| ResNet-152 | 16,232,192 (1,183,990) | **15,048,232 (30)** | **15,048,232 (30)** | 15,048,202 |
| BERT | 28,151,040 (780,480) | **27,569,664 (199,104)** | **27,569,664 (199,104)** | 27,370,560 |

**New Timing Results:**
These are approximate timing results based on my sandbox code. I anticipate the true timing should resemble these numbers. All values are in seconds.

Two processes:
| Model | Run Time - Greedy | Run Time - Sorted-Greedy | Run Time - Multifit |
| --- | --- | --- | --- |
| ResNet-50 | 0.00014 | 0.00014 | 0.00069 |
| ResNet-152 | 0.00036 | 0.00041 | 0.00189 |
| BERT | 0.00015 | 0.00016 | 0.00122 |

Four processes:
| Model | Run Time - Greedy | Run Time - Sorted-Greedy | Run Time - Multifit |
| --- | --- | --- | --- |
| ResNet-50 | 0.00015 | 0.00015 | 0.00095 |
| ResNet-152 | 0.00041 | 0.00042 | 0.00273 |
| BERT | 0.00020 | 0.00019 | 0.00228 |

**Upshot:**
I want to credit the insight from @pritamdamania that `n` in our case should not be too prohibitive in our algorithm design. From the run times, it is certainly feasible to run both greedy and sorted-greedy and take the better partitioning or to run the new algorithm even if may not be able to find a partitioning better than the sorted-greedy algorithm. Either way, the sorted-greedy algorithm appears to be a strong single candidate.